refactor(tops): entschlacken von topsscreen über viewmodel helper
