+++ src/renderer/views/TopsScreen.js
@@ -7,6 +7,17 @@ import { TopsQuicklane } from "../tops/components/TopsQuicklane.js";
 import { TopsWorkbench } from "../tops/components/TopsWorkbench.js";
 import { TopsList } from "../tops/components/TopsList.js";
 import { TopsHeader } from "../tops/components/TopsHeader.js";
+import {
+  buildHeaderContext,
+  buildListItemsFromState,
+  editorFromTop,
+  buildPatchFromDraft,
+  canCreateChildFromState,
+  canDeleteFromState,
+  canMoveFromState,
+  shouldShowWorkbench,
+  buildWorkbenchState,
+} from "../tops/viewmodel/TopsScreenViewModel.js";
 
 const TOPS_V2_STYLE_TAG = "bbm-tops-v2-styles";
 
@@ -211,95 +222,23 @@ export default class TopsScreen {
     this._mountQuicklaneIntoHeader();
   }
 
-  _buildHeaderContext(state) {
-    const m = state?.meetingMeta || null;
-    const parts = [];
-    const projectLine = String(this.router?.context?.projectLabel || "").trim();
-    if (projectLine) parts.push(projectLine);
-
-    const meetingNo = String(m?.meeting_number ?? m?.meetingNumber ?? "").trim();
-    const meetingDate = String(m?.meeting_date ?? m?.meetingDate ?? "").trim();
-    const meetingKeyword = String(m?.keyword || m?.meeting_keyword || "").trim();
-    const meetingLine = [meetingNo, meetingDate].filter(Boolean).join(" - ");
-    if (meetingLine) parts.push(meetingLine);
-    if (meetingKeyword) parts.push(meetingKeyword);
-
-    if (!parts.length) return state?.meetingId ? "Protokoll" : "kein Protokoll aktiv";
-    return parts.join(" | ");
-  }
-
   _syncHeaderState() {
     if (!(this.header instanceof TopsHeader)) return;
     const state = this.store.getState();
     this.header.update({
       title: "Protokoll",
-      context: this._buildHeaderContext(state),
+      context: buildHeaderContext(state, {
+        projectLabel: this.router?.context?.projectLabel || "",
+      }),
       isReadOnly: !!state.isReadOnly,
       canEndMeeting: !!state.meetingId,
       isBusy: !!state.isLoading,
     });
   }
 
-  _buildListItemsFromState() {
-    const state = this.store.getState();
-    const tops = Array.isArray(state.tops) ? state.tops : [];
-    const selectedId = state.selectedTopId;
-    const selectedTop = getSelectedTop(state);
-    const movingTop = state.isMoveMode ? selectedTop : null;
-    const rows = [];
-
-    for (const top of tops) {
-      const due = (top?.due_date || top?.dueDate || "").toString().trim();
-      const status = (top?.status || "").toString().trim();
-      const responsible = (top?.responsible_label || top?.responsibleLabel || "").toString().trim();
-      const meta = [];
-      if (due) meta.push(`Fertig bis: ${due}`);
-      if (status) meta.push(`Status: ${status}`);
-      if (responsible) meta.push(`Verantw.: ${responsible}`);
-
-      let isMoveTarget = null;
-      if (state.isMoveMode && movingTop) {
-        isMoveTarget = this._isAllowedMoveTarget(top, movingTop);
-      }
-
-      rows.push({
-        id: top?.id,
-        level: Number(top?.level) || 1,
-        title: String(top?.title || ""),
-        number: `${top?.displayNumber ?? top?.number ?? ""}.`,
-        preview: String(top?.longtext || ""),
-        meta,
-        isSelected: String(top?.id) === String(selectedId ?? ""),
-        isMoveMode: !!state.isMoveMode,
-        isMoveTarget,
-        raw: top,
-      });
-    }
-    return rows;
-  }
-
   _syncListState() {
     if (!(this.topsList instanceof TopsList)) return;
-    this.topsList.setItems(this._buildListItemsFromState());
-  }
-
-  _hasChildren(topId) {
-    const state = this.store.getState();
-    const tops = Array.isArray(state.tops) ? state.tops : [];
-    const key = String(topId ?? "");
-    return tops.some((t) => String(t?.parent_top_id ?? "") === key);
-  }
-
-  _isBlue(top) {
-    return Number(top?.is_carried_over) !== 1;
-  }
-
-  _isAllowedMoveTarget(target, movingTop) {
-    if (!movingTop || !target) return false;
-    if (String(target.id) === String(movingTop.id)) return false;
-    const tl = Number(target.level);
-    if (!Number.isFinite(tl)) return false;
-    return tl >= 1 && tl <= 3;
+    this.topsList.setItems(buildListItemsFromState(this.store.getState()));
   }
 
   async _handleListRowClick(item) {
@@ -321,110 +260,14 @@ export default class TopsScreen {
     }
 
     this.commands.selectTop(top.id);
-    this.commands.updateDraft(this._editorFromTop(top));
+    this.commands.updateDraft(editorFromTop(top));
     this._syncScreenState();
-  }
-
-  _shouldShowWorkbench(state) {
-    const selectedTop = getSelectedTop(state);
-    const hasSelection = !!selectedTop;
-    const hasMeeting = !!state?.meetingId;
-    if (hasSelection) return true;
-    if (!hasMeeting) return false;
-    if (state?.isReadOnly) return false;
-    return true;
-  }
-
-  _editorFromTop(top) {
-    if (!top) {
-      return {
-        title: "",
-        longtext: "",
-        due_date: null,
-        status: "-",
-        responsible_label: "",
-        is_important: 0,
-        is_hidden: 0,
-        is_decision: 0,
-      };
-    }
-    return {
-      title: String(top.title || ""),
-      longtext: String(top.longtext || ""),
-      due_date: top.due_date || top.dueDate || null,
-      status: String(top.status || "-"),
-      responsible_label: String(top.responsible_label || top.responsibleLabel || ""),
-      is_important: Number(top.is_important) === 1 ? 1 : 0,
-      is_hidden: Number(top.is_hidden) === 1 ? 1 : 0,
-      is_decision: Number(top.is_decision) === 1 ? 1 : 0,
-    };
-  }
-
-  _buildPatchFromDraft(selectedTop, draft) {
-    if (!selectedTop || !draft) return {};
-    const patch = {};
-
-    const title = String(draft.title || "");
-    if (title !== String(selectedTop.title || "")) patch.title = title;
-
-    const longtext = String(draft.longtext || "");
-    if (longtext !== String(selectedTop.longtext || "")) patch.longtext = longtext;
-
-    const dueDate = (draft.due_date || null) || null;
-    const selectedDue = (selectedTop.due_date || selectedTop.dueDate || null) || null;
-    if (dueDate !== selectedDue) patch.due_date = dueDate;
-
-    const status = String(draft.status || "-");
-    if (status !== String(selectedTop.status || "-")) patch.status = status;
-
-    const responsibleLabel = String(draft.responsible_label || "");
-    if (responsibleLabel !== String(selectedTop.responsible_label || selectedTop.responsibleLabel || "")) {
-      patch.responsible_label = responsibleLabel;
-    }
-
-    for (const k of ["is_important", "is_hidden", "is_decision"]) {
-      const nextVal = Number(draft[k]) === 1 ? 1 : 0;
-      const curVal = Number(selectedTop[k]) === 1 ? 1 : 0;
-      if (nextVal !== curVal) patch[k] = nextVal;
-    }
-
-    return patch;
-  }
-
-  _canCreateChildFromState(state, selectedTop) {
-    if (state.isReadOnly || !selectedTop) return false;
-    const level = Number(selectedTop.level);
-    return Number.isFinite(level) && level < 4;
-  }
-
-  _canDeleteFromState(state, selectedTop) {
-    if (state.isReadOnly || !selectedTop) return false;
-    if (!this._isBlue(selectedTop)) return false;
-    if (this._hasChildren(selectedTop.id)) return false;
-    return true;
-  }
-
-  _canMoveFromState(state, selectedTop) {
-    if (state.isReadOnly || !selectedTop) return false;
-    if (!this._isBlue(selectedTop)) return false;
-    if (this._hasChildren(selectedTop.id)) return false;
-    return true;
   }
 
   _syncWorkbenchState() {
     if (!(this.workbench instanceof TopsWorkbench)) return;
     const state = this.store.getState();
-    const selectedTop = getSelectedTop(state);
-    this.workbench.setState({
-      editor: state.editor || this._editorFromTop(selectedTop),
-      isReadOnly: !!state.isReadOnly,
-      hasSelection: !!selectedTop,
-      isMoveMode: !!state.isMoveMode,
-      canSave: !!selectedTop,
-      canDelete: this._canDeleteFromState(state, selectedTop),
-      canMove: this._canMoveFromState(state, selectedTop),
-      canCreateChild: this._canCreateChildFromState(state, selectedTop),
-    });
+    this.workbench.setState(buildWorkbenchState(state));
   }
 
   _syncScreenState() {
@@ -435,7 +278,7 @@ export default class TopsScreen {
     this._syncWorkbenchState();
 
     if (this.editArea) {
-      this.editArea.dataset.bbmWorkbenchVisible = this._shouldShowWorkbench(state) ? "true" : "false";
+      this.editArea.dataset.bbmWorkbenchVisible = shouldShowWorkbench(state) ? "true" : "false";
       this.editArea.dataset.bbmHasSelection = hasSelection(state) ? "true" : "false";
     }
   }
@@ -449,7 +292,7 @@ export default class TopsScreen {
     const state = this.store.getState();
     const selectedTop = getSelectedTop(state);
     if (!selectedTop) return;
-    const patch = this._buildPatchFromDraft(selectedTop, state.editor || {});
+    const patch = buildPatchFromDraft(selectedTop, state.editor || {});
     if (!Object.keys(patch).length) return;
     await this.commands.saveDraft(patch);
     await this._reloadTops({ keepSelection: true, selectTopId: selectedTop.id });
@@ -459,7 +302,7 @@ export default class TopsScreen {
   async _handleWorkbenchDelete() {
     const state = this.store.getState();
     const selectedTop = getSelectedTop(state);
-    if (!this._canDeleteFromState(state, selectedTop)) return;
+    if (!canDeleteFromState(state, selectedTop)) return;
 
     await this.commands.saveDraft({ is_hidden: 1 });
     await this.commands.deleteSelectedTop();
@@ -470,7 +313,7 @@ export default class TopsScreen {
   async _handleWorkbenchToggleMove() {
     const state = this.store.getState();
     const selectedTop = getSelectedTop(state);
-    if (!state.isMoveMode && !this._canMoveFromState(state, selectedTop)) return;
+    if (!state.isMoveMode && !canMoveFromState(state, selectedTop)) return;
     this.commands.toggleMoveMode();
     this._syncScreenState();
   }
@@ -494,7 +337,7 @@ export default class TopsScreen {
   async _handleWorkbenchCreateChild() {
     const state = this.store.getState();
     const selectedTop = getSelectedTop(state);
-    if (!this._canCreateChildFromState(state, selectedTop)) return;
+    if (!canCreateChildFromState(state, selectedTop)) return;
 
     const level = Number(selectedTop.level) + 1;
     if (!Number.isFinite(level) || level > 4) return;
@@ -529,7 +372,7 @@ export default class TopsScreen {
 
     this.commands.selectTop(nextSelectedId);
     const selectedTop = tops.find((t) => String(t?.id) === String(nextSelectedId ?? "")) || null;
-    this.commands.updateDraft(this._editorFromTop(selectedTop));
+    this.commands.updateDraft(editorFromTop(selectedTop));
     this.commands.toggleMoveMode(false);
     this.store.setState({
       meetingMeta: res?.meeting || null,
